Lift loading state so resizing can be disabled
